refactor(services/execution): extract tests via #[path]
Phase 0 PR-4 of the architecture decoupling roadmap (#1082, #1083,
#1084). Same mechanical pattern: move services/execution.rs's
`#[cfg(test)] mod tests` block into a separate file via
`#[cfg(test)] #[path = "..."] mod tests;`.

Net effect:
  services/execution.rs:        2438 -> 1186 lines (-1252)
  services/execution_tests.rs:   NEW   1252 lines

Zero logic changes. New file uses `use super::*;` so it sees the
execution module unchanged. No public API surface change.

Verification:
- cargo test -p harness-server --lib -- services::execution:
    27/27 passed (covers all execution::tests::* including
    enqueue_background_prompt_submission_*, resolve_project_*,
    runtime_prompt_serialization_dependency_*)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean

services/execution.rs is still over the U-16 800-line ceiling (1186
lines remaining). Future Phase 1 cohesive splits can target it.
